@@ -11,6 +11,9 @@
 |
 */
 
+use Illuminate\Support\Facades\Route;
+use Modules\Blog\Http\Controllers\BlogController;
+
 Route::prefix('blog')->group(function() {
-    Route::get('/', 'BlogController@index');
+    Route::get('/', [BlogController::class,'index']);
 });
