@@ -11,6 +11,9 @@
 |
 */
 
+use Illuminate\Support\Facades\Route;
+use Modules\User\Http\Controllers\UserController;
+
 Route::prefix('user')->group(function() {
-    Route::get('/', 'UserController@index');
+    Route::get('/', [UserController::class,'index']);
 });
