@@ -2,8 +2,9 @@
 
 namespace Modules\User\Providers;
 
+use Illuminate\Database\Eloquent\Factories\Factory;
 use Illuminate\Support\ServiceProvider;
-use Illuminate\Database\Eloquent\Factory;
+
 
 class UserServiceProvider extends ServiceProvider
 {
@@ -98,7 +99,7 @@ class UserServiceProvider extends ServiceProvider
     public function registerFactories()
     {
         if (! app()->environment('production') && $this->app->runningInConsole()) {
-            app(Factory::class)->load(module_path($this->moduleName, 'Database/factories'));
+           $this->loadFactoriesFrom(module_path($this->moduleName, 'Database/factories'));
         }
     }
 
